+++ gulp/config.js
@@ -37,9 +37,9 @@ module.exports = {
       'news-list',
       'pagination',
       'person-block',
+      'postit',
       'search-results-list',
-      'share',
-      'postit'
+      'share'
     ]
   },