+++ src/javascript/config.js
@@ -1,32 +1,35 @@
 var Paris = window.Paris || {};
 
 Paris.config = {
-  algolia: {
+  "algolia": {
     // The Algolia application ID
-    id: "QGS0I5WCQR",
+    "id": "QGS0I5WCQR",
 
     // This API key should be created on Algolia, and have `search` permission on all the indexes below
-    api_key: "219f93ef781ffa09cdb6803f702cf6f1",
+    "api_key": "219f93ef781ffa09cdb6803f702cf6f1",
 
     // You can add any Algolia index here, to be able to use it in the quick-access module and search-results template
-    indexes: {
-      global: "recette_ParisFront",
-      persons: "recette_Elus"
+    "indexes": {
+      "global": "recette_ParisFront",
+      "persons": "recette_Elus"
     },
 
-    url: {
-      api_popular_searches: "http://r7.paris-fr-api.lestudio.mx/AlgoliaStats/getMostSearchedKeywords"
+    "url": {
+      "api_popular_searches": "http://r7.paris-fr-api.lestudio.mx/AlgoliaStats/getMostSearchedKeywords"
     }
   },
-  cookies: {
-    email: {
-      name: "email"
+  "cookies": {
+    "email": {
+      "name": "email"
     },
-    cnil: {
-      name: "cookies",
-      value: "accepted",
+    "cnil": {
+      "name": "cookies",
+      "value": "accepted",
       // `expires` can be a Number which will be interpreted as days from time of creation or a Date instance
-      expires: 395 // 13 months
+      "expires": 395 // 13 months
+    },
+    "parisconnect": {
+      "name": "pcuid"
     }
   }
 };