+++ src/templates/document/script.js
@@ -4,20 +4,62 @@ var Cookies = require('js-cookie');
 
 var Paris = window.Paris;
 
-$(document).ready(function(){
-  if(Cookies.get('pcuid')) {
-    var pageId = $('body').data('pageid');
-    $.ajax({
-      url: '/postit/' + pageId,
-      type: 'get',
-      success: renderPostIt
-    });
+Paris.document = (function(){
+
+  var defaultOptions = {
+    postitContentUrl: '/postit/${pageId}'
+  };
+
+  function document(selector, userOptions){
+    var $el = $(selector),
+      options = $.extend({}, defaultOptions, userOptions),
+      pageId
+      ;
+
+    function init(){
+      initOptions();
+
+      pageId = $('body').data('pageid');
+
+      if (Cookies.get(Paris.config.cookies.parisconnect.name)) {
+        loadPostit();
+      }
+    }
+
+    function initOptions() {
+      $.each($el.data(), function(key, value){
+        options[key] = value;
+      });
+    }
+
+    function loadPostit(){
+      $.ajax({
+        url: options.postitContentUrl.replace('${pageId}', pageId),
+        type: 'get',
+        success: renderPostit
+      });
+    }
+
+    function renderPostit(data){
+      data.block = data.contenu;
+      var postit = Paris.templates.templatizer['postit']['postit'](data);
+      $('.components').prepend(postit);
+    }
+
+
+    init();
+
+    return $el;
   }
+
+  return function(selector, userOptions){
+    return $(selector).each(function(){
+      document(this, userOptions);
+    });
+  };
+
+})();
+
+$(document).ready(function(){
+  Paris.document('body.document');
 });
-
-
-function renderPostIt(data) {
-  data.block = data.contenu;
-  var postIt = Paris.templates.templatizer['postit']['postit'](data);
-  $('.layout-content.components').prepend(postIt);
-}